rearrange database folder between data and domain layers
