+++ app/src/main/java/com/practicum/playlistmaker_ver2/App.kt
@@ -3,8 +3,6 @@ package com.practicum.playlistmaker_ver2
 import android.app.Application
 import android.content.Context
 import androidx.appcompat.app.AppCompatDelegate
-import androidx.room.Room
-import com.practicum.playlistmaker_ver2.database.LikedTracks.data.LikedTracksDatabase
 import com.practicum.playlistmaker_ver2.di.dataModule
 import com.practicum.playlistmaker_ver2.di.domainModule
 import com.practicum.playlistmaker_ver2.di.presentationModule
