+++ app/src/main/java/com/practicum/playlistmaker_ver2/di/DataModule.kt
@@ -7,18 +7,18 @@ import android.os.Handler
 import androidx.room.Room
 import com.google.gson.Gson
 import com.practicum.playlistmaker_ver2.App
-import com.practicum.playlistmaker_ver2.database.LikedTracks.data.LikedTracksRepositoryImpl
-import com.practicum.playlistmaker_ver2.database.LikedTracks.data.converters.TrackDbConverter
-import com.practicum.playlistmaker_ver2.database.LikedTracks.data.LikedTracksDatabase
-import com.practicum.playlistmaker_ver2.database.LikedTracks.domain.LikedTracksRepository
-import com.practicum.playlistmaker_ver2.database.Playlists.data.PlaylistDatabase
-import com.practicum.playlistmaker_ver2.player.data.repository.ImageRepositoryImpl
-import com.practicum.playlistmaker_ver2.player.data.repository.PlayerRepositoryImpl
+import com.practicum.playlistmaker_ver2.player.data.impl.LikedTracksRepositoryImpl
+import com.practicum.playlistmaker_ver2.player.data.converters.TrackDbConverter
+import com.practicum.playlistmaker_ver2.player.data.LikedTracksDatabase
+import com.practicum.playlistmaker_ver2.player.domain.repositories.LikedTracksRepository
+import com.practicum.playlistmaker_ver2.playlist.data.repositories.PlaylistDatabase
+import com.practicum.playlistmaker_ver2.player.data.impl.ImageRepositoryImpl
+import com.practicum.playlistmaker_ver2.player.data.impl.PlayerRepositoryImpl
 import com.practicum.playlistmaker_ver2.player.domain.repositories.ImageRepository
 import com.practicum.playlistmaker_ver2.player.domain.repositories.PlayerRepository
-import com.practicum.playlistmaker_ver2.database.Playlists.data.PlaylistRepositoryImpl
-import com.practicum.playlistmaker_ver2.database.Playlists.domain.PlaylistRepository
-import com.practicum.playlistmaker_ver2.database.TracksInPlaylists.TracksInPlaylistsDatabase
+import com.practicum.playlistmaker_ver2.playlist.data.repositories.PlaylistRepositoryImpl
+import com.practicum.playlistmaker_ver2.playlist.domain.PlaylistRepository
+import com.practicum.playlistmaker_ver2.playlist.data.repositories.TracksInPlaylistsDatabase
 
 import com.practicum.playlistmaker_ver2.search.data.dto.TracksRepositoryImpl
 import com.practicum.playlistmaker_ver2.search.data.network.ITunesApiService
