+++ app/src/main/java/com/practicum/playlistmaker_ver2/mediateka/presentation/TabPlaylistsViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.practicum.playlistmaker_ver2.database.Playlists.domain.PlaylistInteractor
+import com.practicum.playlistmaker_ver2.playlist.domain.interactor.PlaylistInteractor
 import com.practicum.playlistmaker_ver2.playlist.domain.models.PlaylistEntityPresentation
 import kotlinx.coroutines.launch
