+++ app/src/main/java/com/practicum/playlistmaker_ver2/player/data/impl/ImageRepositoryImpl.kt
@@ -1,10 +1,7 @@
-package com.practicum.playlistmaker_ver2.player.data.repository
+package com.practicum.playlistmaker_ver2.player.data.impl
 
 import android.content.Context
-import android.graphics.Bitmap
-import android.graphics.BitmapFactory
 import android.net.Uri
-import android.os.Environment
 import android.widget.Toast
 import androidx.core.net.toUri
 import com.practicum.playlistmaker_ver2.player.domain.repositories.ImageRepository
