+++ app/src/main/java/com/practicum/playlistmaker_ver2/player/domain/repositories/LikedTracksRepository.kt
@@ -1,6 +1,5 @@
-package com.practicum.playlistmaker_ver2.database.LikedTracks.domain
+package com.practicum.playlistmaker_ver2.player.domain.repositories
 
-import com.practicum.playlistmaker_ver2.database.LikedTracks.data.TrackEntity
 import com.practicum.playlistmaker_ver2.player.ui.models.PlayerTrack
 import com.practicum.playlistmaker_ver2.search.domain.models.Track
 import kotlinx.coroutines.flow.Flow
