+++ app/src/main/java/com/practicum/playlistmaker_ver2/playlist/data/repositories/TracksInPlaylistsDao.kt
@@ -1,11 +1,11 @@
-package com.practicum.playlistmaker_ver2.database.TracksInPlaylists
+package com.practicum.playlistmaker_ver2.playlist.data.repositories
 
 import androidx.room.Dao
 import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.practicum.playlistmaker_ver2.database.LikedTracks.data.TrackEntity
+import com.practicum.playlistmaker_ver2.playlist.data.entities.TrackInPlaylistEntity
 import kotlinx.coroutines.flow.Flow
 
 @Dao
