+++ app/src/main/java/com/practicum/playlistmaker_ver2/playlist/data/repositories/TracksInPlaylistsDatabase.kt
@@ -1,9 +1,10 @@
-package com.practicum.playlistmaker_ver2.database.TracksInPlaylists
+package com.practicum.playlistmaker_ver2.playlist.data.repositories
 
 import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
+import com.practicum.playlistmaker_ver2.playlist.data.entities.TrackInPlaylistEntity
 
 @Database(
     version = 1, entities = [TrackInPlaylistEntity::class]
