+++ app/src/main/java/com/practicum/playlistmaker_ver2/playlist/domain/implimentation/PlaylistInteractorImpl.kt
@@ -1,6 +1,8 @@
-package com.practicum.playlistmaker_ver2.database.Playlists.domain
+package com.practicum.playlistmaker_ver2.playlist.domain.implimentation
 
 import android.net.Uri
+import com.practicum.playlistmaker_ver2.playlist.domain.interactor.PlaylistInteractor
+import com.practicum.playlistmaker_ver2.playlist.domain.PlaylistRepository
 import com.practicum.playlistmaker_ver2.player.ui.models.PlayerTrack
 import com.practicum.playlistmaker_ver2.playlist.domain.models.PlaylistEntityPresentation
 import kotlinx.coroutines.flow.Flow
