+++ app/src/main/java/com/practicum/playlistmaker_ver2/playlist/presentation/PlaylistViewModel.kt
@@ -7,7 +7,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.practicum.playlistmaker_ver2.mediateka.domain.api.ImageInteractor
-import com.practicum.playlistmaker_ver2.database.Playlists.domain.PlaylistInteractor
+import com.practicum.playlistmaker_ver2.playlist.domain.interactor.PlaylistInteractor
 import com.practicum.playlistmaker_ver2.utils.Constants
 import kotlinx.coroutines.launch
